fully working with two rigid bodies in our OptiTrack system

+++ uvrpn/Assets/Plugins/uvrpn.cs
@@ -18,7 +18,7 @@ namespace uvrpn {
 	[StructLayout(LayoutKind.Sequential)]
 	public struct Sensor
 	{
-		public float posX, posY, posZ;
+		public double posX, posY, posZ;
 		public double quatX,quatY,quatZ,quatW;
 
 	}
